requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.3
